chore: remove old RecyclerAssetList
