--- src/components/asset-list/RecyclerAssetList/index.js
@@ -1,1 +1,0 @@
-export { default } from './RainbowRecyclerAssetList';
